add features in modal
